// cgocore-ui/src/app/features/models/loading/views/basic/DatedFlightUldBasicView.ts
/**
 * auto generated basic view bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class DatedFlightUldBasicView {

id: number;
selected: boolean;
canDelete: boolean;
datedFlightCompanyIataCode: string;
datedFlightFlightNumber: string;
datedFlightDepartureDate: string;
uldTypeCode: string;
uldCode: string;
uldOwnerCode: string;
grossWeight: number;
grossWeightUnitOfMeasureCode: string;

/* Specific Code Start */
/* Specific Code End */
}