// cgocore-ui/src/app/features/models/loading/sortings/DatedFlightUldSorting.ts
/**
 * auto generated sorting ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class DatedFlightUldSorting {

datedFlightCompanyIataCodeOrderType: string;
datedFlightFlightNumberOrderType: string;
datedFlightDepartureDateOrderType: string;
uldTypeCodeOrderType: string;
uldCodeOrderType: string;
uldOwnerCodeOrderType: string;
grossWeightOrderType: string;
grossWeightUnitOfMeasureCodeOrderType: string;

/* Specific Code Start */
/* Specific Code End */
}